+++ lib/MLD.h
@@ -42,23 +42,23 @@
 #include <stdio.h>
 #endif /* __tinyos__ */
 
-	uint8_t MLD_getObjectWithIndex(uint8_t ndex, uint8_t buffer[], uint8_t length, ml_object_t *object) ATTR_TINYOS_AT_C;
+	uint8_t MLD_getObjectWithIndex(uint8_t ndex, uint8_t buffer[], uint16_t length, ml_object_t *object) ATTR_TINYOS_AT_C;
 
-	uint8_t MLD_findObjectWithParameters(uint32_t type, uint8_t* subject, int32_t* value, uint8_t buffer[], uint8_t length, ml_object_t* object) ATTR_TINYOS_AT_C;
-	uint8_t MLD_findOSV(uint32_t type, uint8_t subject, int32_t value, uint8_t buffer[], uint8_t length, ml_object_t* object) ATTR_TINYOS_AT_C;
-	uint8_t MLD_findOV(uint32_t type, int32_t value, uint8_t buffer[], uint8_t length, ml_object_t* object) ATTR_TINYOS_AT_C;
-	uint8_t MLD_findOS(uint32_t type, uint8_t subject, uint8_t buffer[], uint8_t length, ml_object_t* object) ATTR_TINYOS_AT_C;
-	uint8_t MLD_findO(uint32_t type, uint8_t buffer[], uint8_t length, ml_object_t* object) ATTR_TINYOS_AT_C;
+	uint8_t MLD_findObjectWithParameters(uint32_t type, uint8_t* subject, int32_t* value, uint8_t buffer[], uint16_t length, ml_object_t* object) ATTR_TINYOS_AT_C;
+	uint8_t MLD_findOSV(uint32_t type, uint8_t subject, int32_t value, uint8_t buffer[], uint16_t length, ml_object_t* object) ATTR_TINYOS_AT_C;
+	uint8_t MLD_findOV(uint32_t type, int32_t value, uint8_t buffer[], uint16_t length, ml_object_t* object) ATTR_TINYOS_AT_C;
+	uint8_t MLD_findOS(uint32_t type, uint8_t subject, uint8_t buffer[], uint16_t length, ml_object_t* object) ATTR_TINYOS_AT_C;
+	uint8_t MLD_findO(uint32_t type, uint8_t buffer[], uint16_t length, ml_object_t* object) ATTR_TINYOS_AT_C;
 
-	uint8_t MLD_getObjectCount(uint8_t buffer[], uint8_t length) ATTR_TINYOS_AT_C;
+	uint8_t MLD_getObjectCount(uint8_t buffer[], uint16_t length) ATTR_TINYOS_AT_C;
 
 	/* Protected */
 
 	uint8_t pMLD_readUnsignedByte(uint8_t* bstart) ATTR_TINYOS_AT_C;
 	int32_t pMLD_readInteger(uint8_t num, uint8_t* bstart) ATTR_TINYOS_AT_C;
-	uint8_t pMLD_getObjectAt(uint8_t offset, uint8_t buffer[], uint8_t length, ml_object_t* object) ATTR_TINYOS_AT_C;
-	uint8_t pMLD_getTypeAt(uint8_t offset, uint8_t buffer[], uint8_t length, uint32_t* type) ATTR_TINYOS_AT_C;
-	uint8_t pMLD_getNextOffset(uint8_t offset, uint8_t buffer[], uint8_t length) ATTR_TINYOS_AT_C;
-	uint8_t pMLD_getOffsetOf(uint8_t buffer[], uint8_t length, uint8_t number) ATTR_TINYOS_AT_C;
+	uint16_t pMLD_getObjectAt(uint16_t offset, uint8_t buffer[], uint16_t length, ml_object_t* object) ATTR_TINYOS_AT_C;
+	uint8_t pMLD_getTypeAt(uint16_t offset, uint8_t buffer[], uint16_t length, uint32_t* type) ATTR_TINYOS_AT_C;
+	uint16_t pMLD_getNextOffset(uint16_t offset, uint8_t buffer[], uint16_t length) ATTR_TINYOS_AT_C;
+	uint16_t pMLD_getOffsetOf(uint8_t buffer[], uint16_t length, uint8_t number) ATTR_TINYOS_AT_C;
 
 #endif /* MLD_H_ */
